refactor: add region view
