+++ src/pages/MapPage/components/LineBreak.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function LineBreak() {
+    return <hr className="h-[1.5px] bg-gray-200 border-0 rounded-2xl" />;
+}
+
+export default LineBreak;
